requirements.txt
google-generativeai
pandas
python-dotenv
streamlit
tabulate
plotly
openpyxl
fpdf2